Renaming smallfull package to ixbrl
