+++ src/main/java/uk/gov/companieshouse/api/accounts/model/ixbrl/balancesheet/CalledUpSharedCapitalNotPaid.java
@@ -1,4 +1,4 @@
-package uk.gov.companieshouse.api.accounts.model.smallfull.balancesheet;
+package uk.gov.companieshouse.api.accounts.model.ixbrl.balancesheet;
 
 public class CalledUpSharedCapitalNotPaid {
     private int currentAmount;
